refactor(cfg): homogenous design with cli
