+++ cmd/internal/root/process_obtain.go
@@ -2,160 +2,107 @@ package root
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"net/http"
-	"sync"
-	"time"
 
 	"github.com/go-acme/lego/v5/certcrypto"
-	"github.com/go-acme/lego/v5/challenge"
-	"github.com/go-acme/lego/v5/cmd/internal"
+	"github.com/go-acme/lego/v5/certificate"
 	"github.com/go-acme/lego/v5/cmd/internal/configuration"
 	"github.com/go-acme/lego/v5/cmd/internal/storage"
 	"github.com/go-acme/lego/v5/lego"
-	"github.com/go-acme/lego/v5/registration"
 )
 
-func obtain(ctx context.Context, cfg *configuration.Configuration) error {
-	networkStack := getNetworkStack(cfg)
+func obtain(ctx context.Context, lazySetup lzSetUp, certConfig *configuration.Certificate, certsStorage *storage.CertificatesStorage) error {
+	client, err := lazySetup()
+	if err != nil {
+		return fmt.Errorf("set up client: %w", err)
+	}
 
-	store := storage.New(cfg.Storage)
+	if certConfig.CSR != "" {
+		return obtainForCSR(ctx, client, certConfig, certsStorage)
+	}
 
-	for accountID, challengesInfo := range createCertificatesMapping(cfg) {
-		accountConfig := cfg.Accounts[accountID]
+	return obtainForDomains(ctx, client, certConfig, certsStorage)
+}
 
-		keyType, err := certcrypto.ToKeyType(accountConfig.KeyType)
-		if err != nil {
-			return err
-		}
+func obtainForDomains(ctx context.Context, client *lego.Client, certConfig *configuration.Certificate, certsStorage *storage.CertificatesStorage) error {
+	keyType, err := certcrypto.ToKeyType(certConfig.KeyType)
+	if err != nil {
+		return fmt.Errorf("get the key type: %w", err)
+	}
 
-		serverConfig := configuration.GetServerConfig(cfg, accountID)
+	request := certificate.ObtainRequest{
+		Domains:                        certConfig.Domains,
+		KeyType:                        keyType,
+		MustStaple:                     certConfig.MustStaple,
+		NotBefore:                      certConfig.NotBefore,
+		NotAfter:                       certConfig.NotAfter,
+		Bundle:                         !certConfig.NoBundle,
+		PreferredChain:                 certConfig.PreferredChain,
+		EnableCommonName:               certConfig.EnableCommonName,
+		Profile:                        certConfig.Profile,
+		AlwaysDeactivateAuthorizations: certConfig.AlwaysDeactivateAuthorizations,
+	}
 
-		account, err := store.Account.Get(serverConfig.URL, keyType, accountConfig.Email, accountID)
-		if err != nil {
-			return err
-		}
+	// NOTE(ldez): I didn't add an option to set a private key as the file.
+	// I didn't find a use case for it when using the file configuration.
+	// Maybe this can be added in the future.
 
-		lazyClient := sync.OnceValues(func() (*lego.Client, error) {
-			client, errC := lego.NewClient(newClientConfig(serverConfig, account, cfg.UserAgent))
-			if errC != nil {
-				return nil, errC
-			}
+	certRes, err := client.Certificate.Obtain(ctx, request)
+	if err != nil {
+		return err
+	}
 
-			if client.GetServerMetadata().ExternalAccountRequired && accountConfig.ExternalAccountBinding == nil {
-				return nil, errors.New("server requires External Account Binding (EAB)")
-			}
-
-			return client, nil
-		})
-
-		err = handleRegistration(ctx, lazyClient, accountConfig, store.Account, account)
-		if err != nil {
-			return fmt.Errorf("registration: %w", err)
-		}
-
-		for challengeID, certIDs := range challengesInfo {
-			chlgConfig := cfg.Challenges[challengeID]
-
-			lazySetup := sync.OnceValues(func() (*lego.Client, error) {
-				client, errC := lazyClient()
-				if errC != nil {
-					return nil, fmt.Errorf("set up client: %w", errC)
-				}
-
-				client.Challenge.RemoveAll()
-
-				errC = setupChallenges(client, chlgConfig, networkStack)
-				if errC != nil {
-					return nil, fmt.Errorf("setup challenges: %w", errC)
-				}
-
-				return client, nil
-			})
-
-			for _, certID := range certIDs {
-				certConfig := cfg.Certificates[certID]
-
-				// Renew
-				if store.Certificate.ExistsFile(certID, storage.ExtResource) {
-					err = renewCertificate(ctx, lazyClient, certID, certConfig, store.Certificate)
-					if err != nil {
-						return err
-					}
-
-					continue
-				}
-
-				// Run
-				err := runCertificate(ctx, lazySetup, certConfig, store.Certificate)
-				if err != nil {
-					return err
-				}
-			}
-		}
+	err = certsStorage.Save(
+		&storage.Certificate{
+			Resource: certRes,
+			Origin:   storage.OriginConfiguration,
+		},
+		&storage.SaveOptions{PEM: true},
+	)
+	if err != nil {
+		return fmt.Errorf("could not save the resource: %w", err)
 	}
 
 	return nil
 }
 
-// createCertificatesMapping creates a mapping of account -> challenge -> certificate IDs.
-func createCertificatesMapping(cfg *configuration.Configuration) map[string]map[string][]string {
-	// Accounts -> Challenges -> Certificates
-	certsMappings := make(map[string]map[string][]string)
-
-	for certID, certDesc := range cfg.Certificates {
-		if _, ok := certsMappings[certDesc.Account]; !ok {
-			certsMappings[certDesc.Account] = make(map[string][]string)
-		}
-
-		certsMappings[certDesc.Account][certDesc.Challenge] = append(certsMappings[certDesc.Account][certDesc.Challenge], certID)
+func obtainForCSR(ctx context.Context, client *lego.Client, certConfig *configuration.Certificate, certsStorage *storage.CertificatesStorage) error {
+	csr, err := storage.ReadCSRFile(certConfig.CSR)
+	if err != nil {
+		return err
 	}
 
-	return certsMappings
-}
-
-func getNetworkStack(cfg *configuration.Configuration) challenge.NetworkStack {
-	switch cfg.NetworkStack {
-	case "ipv4only", "ipv4":
-		return challenge.IPv4Only
-
-	case "ipv6only", "ipv6":
-		return challenge.IPv6Only
-
-	default:
-		return challenge.DualStack
-	}
-}
-
-func newClientConfig(serverConfig *configuration.Server, account registration.User, ua string) *lego.Config {
-	config := lego.NewConfig(account)
-	config.CADirURL = serverConfig.URL
-	config.UserAgent = ua
-	config.Certificate = lego.CertificateConfig{}
-
-	if serverConfig.OverallRequestLimit > 0 {
-		config.Certificate.OverallRequestLimit = serverConfig.OverallRequestLimit
+	// obtain a certificate for this CSR
+	request := certificate.ObtainForCSRRequest{
+		CSR:                            csr,
+		NotBefore:                      certConfig.NotBefore,
+		NotAfter:                       certConfig.NotAfter,
+		Bundle:                         !certConfig.NoBundle,
+		PreferredChain:                 certConfig.PreferredChain,
+		EnableCommonName:               certConfig.EnableCommonName,
+		Profile:                        certConfig.Profile,
+		AlwaysDeactivateAuthorizations: certConfig.AlwaysDeactivateAuthorizations,
 	}
 
-	if serverConfig.CertTimeout > 0 {
-		config.Certificate.Timeout = time.Duration(serverConfig.CertTimeout) * time.Second
+	// NOTE(ldez): I didn't add an option to set a private key as the file.
+	// I didn't find a use case for it when using the file configuration.
+	// Maybe this can be added in the future.
+
+	certRes, err := client.Certificate.ObtainForCSR(ctx, request)
+	if err != nil {
+		return err
 	}
 
-	if serverConfig.HTTPTimeout > 0 {
-		config.HTTPClient.Timeout = time.Duration(serverConfig.HTTPTimeout) * time.Second
+	err = certsStorage.Save(
+		&storage.Certificate{
+			Resource: certRes,
+			Origin:   storage.OriginConfiguration,
+		},
+		&storage.SaveOptions{PEM: true},
+	)
+	if err != nil {
+		return fmt.Errorf("could not save the resource: %w", err)
 	}
 
-	if serverConfig.TLSSkipVerify {
-		defaultTransport, ok := config.HTTPClient.Transport.(*http.Transport)
-		if ok { // This is always true because the default client used by the CLI defined the transport.
-			tr := defaultTransport.Clone()
-			tr.TLSClientConfig.InsecureSkipVerify = true
-			config.HTTPClient.Transport = tr
-		}
-	}
-
-	config.HTTPClient = internal.NewRetryableClient(config.HTTPClient)
-
-	return config
+	return nil
 }
